getOffsetCoordinates util file
