+++ packages/elements/src/utils/getOffsetCoordinates.ts
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2021 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * External dependencies
+ */
+import { PAGE_HEIGHT, PAGE_WIDTH } from '@googleforcreators/units';
+
+/**
+ * Gets x, y values for cloned element, ensuring it's not added out of the page.
+ *
+ * @param originX Original X.
+ * @param originY Original Y.
+ * @return Coordinates.
+ */
+export function getOffsetCoordinates(originX: number, originY: number) {
+  const placementDiff = 30;
+  const allowedBorderDistance = 20;
+  const x = originX + placementDiff;
+  const y = originY + placementDiff;
+  return {
+    x: PAGE_WIDTH - x > allowedBorderDistance ? x : placementDiff,
+    y: PAGE_HEIGHT - y > allowedBorderDistance ? y : placementDiff,
+  };
+}
